workbench: Load local settings by environment variable
When installing Shoop in non-develop mode, the local_settings.py file
will not be copied to site-packages (or even if it would, it would get
outdated quite easily).  To allow specifying local settings then, read
environment variable LOCAL_SETTINGS_FILE and if it is set, load local
settings from the file it points to.

Refs SHOOP-994 / SHOOP-1027