Add shoop_makemessages management command
Add a management command "shoop_makemessages" as a helper for Shoop
projects.  Makes it easier to extract the messages, since the command to
remember is much shorter.

Runs Django's makemessages for "django" and "djangojs" domains with
better defaults for Shoop projects (ignores, file extensions, etc.).

* Accepts same options as bare "makemessages".
* Defaults to --no-location and --no-obsolete.
  - These can be overridden with added --include-* options.

Note: Should not be ran from the project root but from the shoop
directory.

Refs SHOOP-1829/SHOOP-1886