OrderCreator: Make request argument optional
Address a long standing TODO item of removing OrderCreator.request.

The only thing holding us back to not remove request from OrderCreator
was that it is passed as argument to order_creator_finished signal, but
now that the signal was moved from Front to Core which already is a API
breaking change, there is a good chance to remove the request
requirement from OrderCreator.

* Make OrderCreator initializable without a request.
* Document the OrderCreator initializer.
* Deprecate creating OrderCreator with a request.
* Add ip_address to OrderSource.
* Initialize ip_address of BaseBasket from request.
* Remove request parameter from the order_creator_finished signal.
* Make pricing_context optional for update_order_line_from_product.
* Document pricing_context parameter of update_order_line_from_product.

Refs SHOOP-872, SHOOP-2114