admin: Fix generated_resources.txt
In commit e657ce669 some new resources were added for
gulp (i.e. `contact-group.js*`), but they were not added to
generated_resources.txt. This causes the build cache to not function
properly, since it does not know which of the files are source files and
which are generated.  Add the missing resources to the
`generated_resources.txt` file (and sort it).

Refs SHOOP-1981