Xtheme: Enable global placeholder type
Enable a ``global`` parameter for Xtheme placeholders which would store
a site-wite placeholder configuration.

This is done by overriding the placeholder's view name as
``_XthemeGlobalView`` and should work as expected when editing,
rendering, etc.

Currently, this only applies to global placeholders that are within the
same template (but might be included in other templates), so if two
placeholders are marked as ``global`` but are not in the same template
they will not share the same settings.

Refs SHOOP-2219