CORE: Overwrite empty language fields with default value
When a user completes a form that includes translatable fields (e.g.
Product) and only fill out the default language portion (e.g. English)
the form will fill out the other language fields with that of the
english. This fix is meant to address limitations in django-parlers own
fallback mechanism.

This commit also includes a data migration that will perform the same
functionality on existing translation models.

Refs SH-90