Core: Price rounding change
* Round priceful taxless and taxful prices into two decimals
* Round taxes to two decimals when getting order line tax
amount.
* Limit source line base_unit_price and discount_amount to
9 decimals like it is in order line
* Limit priceful price to 9 decimals
* Remove uneccessary rounding while calculating order totals

One option would have been to round priceful price to two decimals,
but it will cause problems when calculating other price components
like discount rate since all the other price components would be
with full precision.

All spots with rounding to two decimals can be changed to round
based on currency decimals later.

See the extra test for refunds which will not be solved by
this PR in case that is a problem. I think we shouldn't refund
with higher precision that the currency allows.

Refs SH-129