Fix issue with stacking filters
In old implementation, whenever the `get_queryset(...)` returned an
empty queryset or `None`, the old queryset was used.

This caused the following prolem:
* Product A is has manufacturer Y defined
* Product B is a variation with Colors Red and Green
* Now filter to show manufacturer Y and color Red, you will see Product A.

The reason this happened was that the queryset handling
in `get_product_queryset` ignored all the querysets that resulted empty
(alongside the `None` values it should ignore). And when doing so, it
did fallback to the previous queryset, which still had the
previous filters' results in it thus could never end in a situtiation where
nothing was shown.

Refs SH-284