front,themes: Add missing owl.video.play.png static file
The CSS in Shuup Front and Classic Gray Theme is using Owl Carousel
which has an url reference to file owl.video.play.png, but that file was
not copied to static files.  Copy it too, so that the CSS doesn't have
any broken url references anymore.

This is important, because if the project is configured to use Django's
ManifestStaticFilesStorage, then "manage.py collectstatic" will break,
if the CSS has broken url references.