front: Improve CheckoutPhaseStorage API
Remove `storage.has_keys(keys)` and add `storage.keys()`.

Remove the `has_keys` method, which was never released (it was added
after 1.1) and is redundant, since there already is a method called
`has_any` which does the same thing.

The `keys` method makes the storage easier to introspect, since it is
now easier to see the keys stored in the storage.  It is also
intentionally named with the same name as `keys` method of hte `dict`.