Front: fixed copy of address values
When regions module is enabled, the region code field is replaced by another field, breaking the original event listeners on that field.
The fields copy is done in 2 passes for each field. The first copy triggers the fields change event.
In this case, the country field makes the region code field to change, either by loading the regions of the country or
by being replaced by a new field. When this occurs, we do a new copy pass as now the region code should contain the correct region options.

Refs POS-2490