Front: Improve  company registration logic
The problem was that when going to Company Information, it was throwing a full 404 (not the django one),
The button was being shown despite Shop not allowing company registration, which was causing the above mentioned error.
This ensures button is shown if customer is a company contact and if company registration is allowed.
Django 404 error is thrown instead of HttpResponseNotFound.
No refs