Fix setting M2M fields
This breaks Django 1.8 and 1.9 compatibility. Could this
be still compatible. Setting M2M is no something that should
happen very often and a lot of these is in tests which could
be just replaced with add.

I guess the question is that is setting M2M in those places
neccessary and is this originally bad code?