Change timezone in middleware
Activated the timezone in the middleware. This was done by adding the timezone field to the users contact from, so the user can select there own timezone.
Then I added the timezone field to the supplier model so the supplier can specify their timezone. This is the fallback timezone if the user haven't specified one.
The last fallback timezone if the nether the supplier of user has specified a timezone is the timezone specified in setting.