Front: enforce setting the current customer to the basket
When there is a customer and the basket still doesn't have a customer,
set it on the address phase.

There are cases when an anonymous user create a basket (adding a product)
and the checkout phases can't have access to the current basket.customer
as it was never set after the user logs is.

Only set the customer when there is no one set previously.

Refs https://trello.com/c/YnkkHRR5